Seperate the pie chart processing to reuse the same code for both charts
