+++ park-dashboard/src/Components/Alerts3dPieChart/AlertsVsAllPark.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { LoadAlertData } from "./LoadAlertData";
-import Alerts3dPieChartsProgressing from "./Alert3dPieChartsProgressing";
+import PieChartsProgressing from "./Alert3dPieChartsProgressing";
+
 function AlertsVsParks() {
   const alertData = LoadAlertData();
-
-  return <Alerts3dPieChartsProgressing alertData={alertData} />;
+  return <PieChartsProgressing alertData={alertData} />;
 }
 
 export default AlertsVsParks;
